pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "heimdallr-py"
version = "0.0.0.dev1"
authors = [
  { name="Grant Giesbrecht", email="grant.giesbrecht@colorado.edu" },
]
description = "Modular instrument control and lab access via Python"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
]
dependencies = [
	'numpy >= 1.0.0',
	'pyvisa >= 1.0.0',
	'pylogfile >= 0.3.0',
	'h5py >= 3.0.0',
	'colorama >= 0.4.0',
	'zhinst >= 24.4.2',
	'pyfrost-network >= 0.0.0.dev0'
]

[project.urls]
Homepage = "https://github.com/Grant-Giesbrecht/heimdallr"
Issues = "https://github.com/Grant-Giesbrecht/heimdallr/issues"